Update merels.pl
All the merels have been placed. We can get a (legal) move from the player named in argument 1, move the merel he or she gives, check for any new mills, and ask which piece to remove if so, display the board, switch players and then play again, with the updated board and the new player.